Trabajo terminado

+++ app/src/main/java/com/example/landscapecontacts/ContactsDetails.kt
@@ -2,15 +2,14 @@ package com.example.landscapecontacts
 
 
 import android.annotation.SuppressLint
+import android.content.DialogInterface
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageButton
-import android.widget.ImageView
-import android.widget.TextView
+import android.widget.*
+import androidx.appcompat.app.AlertDialog
 import com.bumptech.glide.Glide
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
@@ -26,11 +25,9 @@ class ContactsDetails : Fragment() {
     private lateinit var image: ImageView
     private lateinit var imageUrl: String
     private val db = Firebase.firestore
-    private val main = MainActivity()
     //Instanciamos el interface encargado de comunicar los fragment y el MainActivity
     private lateinit var communicator: Communicator
 
-    @SuppressLint("NotifyDataSetChanged")
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -38,6 +35,7 @@ class ContactsDetails : Fragment() {
         // creamos variable vista, instanciamos la vista del contact_details
         val vista: View = inflater.inflate(R.layout.contacts_details, container, false)
 
+        //Communicator
         communicator = activity as Communicator
 
         // instanciamos los componentes visuales del layout a traves de sus ids
@@ -60,6 +58,7 @@ class ContactsDetails : Fragment() {
         val fragmentLayout: Int = (R.id.frag_3)
         val fragLayOut2: Int = (R.id.frag)
 
+        //Modificar el contacto
         btn.setOnClickListener {
             // usamos bundle para almacenar los datos en "cache" y asi poder usarlos en el otro fragment
             val bundle = Bundle()
@@ -79,9 +78,7 @@ class ContactsDetails : Fragment() {
         //Borramos el contacto
         val btnDelete: ImageButton = vista.findViewById(R.id.delete)
         btnDelete.setOnClickListener {
-            deleteContact(number.text.toString())
-            main.deleteFrag(fragmentManager!!, fragLayOut2)
-            communicator.restartFrag(true)
+            showDialogAlertSimple(number.text.toString(), fragLayOut2, communicator)
         }
 
         // una vez que ya tenemos los valores instanciados en sus respectivas variables, solo resta inflar la vista
@@ -102,6 +99,7 @@ class ContactsDetails : Fragment() {
         }
     }
 
+    //Borra el fragment
     private fun deleteFrag(fragLayOut: Int) {
         val frag = fragmentManager?.findFragmentById(fragLayOut)
         val transaction = fragmentManager?.beginTransaction()
@@ -113,8 +111,23 @@ class ContactsDetails : Fragment() {
         }
     }
 
-    //Borra el documento
-    private fun deleteContact(number: String){
-        db.collection("contacts").document(number).delete()
+    //Mostramos el dialogo
+    fun showDialogAlertSimple(number: String, fragLayOut: Int, communicator: Communicator) {
+        AlertDialog.Builder(context!!)
+            .setTitle(getString(R.string.alert))
+            .setMessage(getString(R.string.delete_contact))
+            .setPositiveButton(android.R.string.ok,
+                DialogInterface.OnClickListener { dialog, which ->
+                    db.collection("contacts").document(number).delete()
+                    deleteFrag(fragLayOut)
+                    communicator.restartFrag(true)
+                })
+            .setNegativeButton(android.R.string.cancel,
+                DialogInterface.OnClickListener { dialog, which ->
+                })
+            //Con esta opcion, quitamos la opcion de tocar en la pantalla gris y que se cierre
+            //Si el valor es false, no se cierra, con true si. true viene por defecto
+            .setCancelable(true)
+            .show()
     }
 }
